fs: merge FileHandleStorage into file_storage TU
