+++ libraries/libstratosphere/include/stratosphere/fs/fs_file_storage.hpp
@@ -14,13 +14,14 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 #pragma once
-#include "fs_common.hpp"
-#include "fs_istorage.hpp"
-#include "fsa/fs_ifile.hpp"
+#include <stratosphere/fs/fs_common.hpp>
+#include <stratosphere/fs/fs_istorage.hpp>
+#include <stratosphere/fs/fsa/fs_ifile.hpp>
+#include <stratosphere/fs/impl/fs_newable.hpp>
 
 namespace ams::fs {
 
-    class FileStorage : public IStorage {
+    class FileStorage : public IStorage, public impl::Newable {
         private:
             static constexpr s64 InvalidSize = -1;
         private:
@@ -42,6 +43,34 @@ namespace ams::fs {
             }
 
             virtual ~FileStorage() { /* ... */ }
+        protected:
+            Result UpdateSize();
+        public:
+            virtual Result Read(s64 offset, void *buffer, size_t size) override;
+            virtual Result Write(s64 offset, const void *buffer, size_t size) override;
+            virtual Result Flush() override;
+            virtual Result GetSize(s64 *out_size) override;
+            virtual Result SetSize(s64 size) override;
+            virtual Result OperateRange(void *dst, size_t dst_size, OperationId op_id, s64 offset, s64 size, const void *src, size_t src_size) override;
+    };
+
+    class FileHandleStorage : public IStorage, public impl::Newable {
+        private:
+            static constexpr s64 InvalidSize = -1;
+        private:
+            FileHandle handle;
+            bool close_file;
+            s64 size;
+            os::Mutex mutex;
+        public:
+            constexpr explicit FileHandleStorage(FileHandle handle, bool close_file) : handle(handle), close_file(close_file), size(InvalidSize), mutex() { /* ... */ }
+            constexpr explicit FileHandleStorage(FileHandle handle) : FileHandleStorage(handle, false) { /* ... */ }
+
+            virtual ~FileHandleStorage() override {
+                if (this->close_file) {
+                    CloseFile(this->handle);
+                }
+             }
         protected:
             Result UpdateSize();
         public:
